adgm-corporate-agent/requirements.txt
streamlit>=1.28.0
python-docx==0.8.11
openai>=1.3.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
numpy>=1.26.0
python-magic==0.4.27
langchain>=0.0.300
pypdf>=3.17.0
python-dotenv>=1.0.0
PyPDF2==3.0.1  
pdfplumber==0.9.0 
